+++ Woofer/Systems/Movement/PlayerMovementComponent.cs
@@ -9,15 +9,15 @@ namespace WooferGame.Systems.Movement
     class PlayerMovementComponent : Component
     {
         [PersistentProperty]
-        public double WalkSpeed { get; set; } = 100;
+        public float WalkSpeed { get; set; } = 100;
         [PersistentProperty]
-        public double MaxWalkSpeed { get; set; } = 100;
+        public float MaxWalkSpeed { get; set; } = 100;
         [PersistentProperty]
-        public double AirborneSpeed { get; set; } = 6;
+        public float AirborneSpeed { get; set; } = 6;
         [PersistentProperty]
-        public double MaxAirborneSpeed { get; set; } = 30;
+        public float MaxAirborneSpeed { get; set; } = 30;
         [PersistentProperty]
-        public double JumpSpeed { get; set; } = 192;
+        public float JumpSpeed { get; set; } = 192;
 
         public double CurrentSpeed => OnGround ? WalkSpeed : AirborneSpeed;
         public double CurrentMaxSpeed => OnGround ? MaxWalkSpeed : MaxAirborneSpeed;
